moving compute from child to parent
